requirements.txt
discord.py
python-dotenv
psutil
Pillow
PyNaCl